# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Katalog Service API",
        "description": "API dokumentasi untuk service Katalog Barang pada Sistem Lelang",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/v1/items": {
            "get": {
                "tags": [
                    "Katalog"
                ],
                "summary": "Ambil semua daftar barang lelang",
                "operationId": "5561acef6e4aef27354515aca3069b72",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Katalog"
                ],
                "summary": "Tambah item baru ke katalog lelang",
                "operationId": "a4fb16eb9970fcb06a5d70ace3026f4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Lukisan Vintage"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lukisan langka tahun 1920"
                                    },
                                    "starting_price": {
                                        "type": "integer",
                                        "example": 5000000
                                    },
                                    "auction_deadline": {
                                        "type": "string",
                                        "example": "2026-07-01 18:00:00"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item berhasil ditambahkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthorized - API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/items/{id}": {
            "get": {
                "tags": [
                    "Katalog"
                ],
                "summary": "Ambil detail spesifik satu barang",
                "operationId": "b7db156ef0b44edcbe062490557a9c94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item tidak ditemukan"
                    },
                    "401": {
                        "description": "Unauthorized - API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/items/filter": {
            "post": {
                "tags": [
                    "Katalog"
                ],
                "summary": "Filter barang berdasarkan kriteria tertentu",
                "operationId": "af898aafc55e5f97fbcf023e5bf5cfc4",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "auction_status": {
                                        "type": "string",
                                        "example": "OPEN"
                                    },
                                    "min_price": {
                                        "type": "integer",
                                        "example": 1000000
                                    },
                                    "max_price": {
                                        "type": "integer",
                                        "example": 10000000
                                    },
                                    "keyword": {
                                        "type": "string",
                                        "example": "lukisan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil filter data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data filtered successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ada item yang sesuai filter"
                    },
                    "401": {
                        "description": "Unauthorized - API Key tidak valid"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Katalog",
            "description": "Katalog"
        }
    ]
}